Abstracting the edit property into a separate file
